@@ -1,71 +1,74 @@
 {
   "name": "@dzangolab/vue-accordion",
-  "version": "1.1.0",
-  "description": "An accordion component for vue.js",
-  "author": {
-    "name": "Olivier Pichon",
-    "email": "op@dzango.com",
-    "url": "https://github.com/opichon"
-  },
-  "scripts": {
-    "build": "vue-cli-service build --target lib --name accordion ./src/components/Accordion/Index.vue",
-    "commitlint": "commitlint --edit $1",
-    "lint": "vue-cli-service lint",
-    "prerelease": "yarn run test",
-    "prepublishOnly": "yarn run lint && yarn run test && yarn run build",
-    "release": "standard-version",
-    "serve": "vue-cli-service serve",
-    "test": "yarn run test:unit",
-    "test:unit": "vue-cli-service test:unit",
-    "prepare": "husky install"
-  },
-  "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2",
-    "@fortawesome/free-solid-svg-icons": "^5.15",
-    "@fortawesome/vue-fontawesome": "^2.0",
-    "bootstrap": "^4.6",
-    "core-js": "^3.9",
-    "vue": "^2.6"
-  },
-  "devDependencies": {
-    "@commitlint/cli": "^12.0",
-    "@commitlint/config-conventional": "^12.0",
-    "@vue/cli-plugin-babel": "^4.5",
-    "@vue/cli-plugin-eslint": "^4.5",
-    "@vue/cli-plugin-unit-jest": "^4.5",
-    "@vue/cli-service": "^4.5",
-    "@vue/test-utils": "1.1",
-    "@vuedoc/md": "^3.2.0",
-    "@vuedoc/parser": "^3.3.0",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "^10.0",
-    "babel-jest": "^26.6",
-    "eslint": "^7.22",
-    "eslint-plugin-import": "^2.22",
-    "eslint-plugin-node": "^11.1",
-    "eslint-plugin-promise": "^4.3",
-    "eslint-plugin-standard": "^5.0",
-    "eslint-plugin-vue": "^7.7",
-    "husky": "5",
-    "node-sass": "^5.0",
-    "sass-loader": "^10.1",
-    "standard-version": "^9.1",
-    "vue-template-compiler": "^2.6"
-  },
-  "bugs": "https://github.com/dzangolab/vue-accordion/issues",
-  "files": [
-    "dist/*",
-    "src/*",
-    "public/*",
-    "*.json",
-    "*.js"
-  ],
-  "homepage": "https://github.com/dzangolab/vue-accordion",
+  "version": "2.0.2",
+  "description": "An accordion component for vue.js (vue3)",
   "keywords": [
     "vuejs",
     "accordion"
   ],
+  "homepage": "https://github.com/dzangolab/vue-accordion",
+  "bugs": "https://github.com/dzangolab/vue-accordion/issues",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/dzangolab/vue-accordion.git"
+  },
   "license": "MIT",
+  "author": {
+    "name": "Dzango Technologies Limited",
+    "email": "info@dzango.com",
+    "url": "https://www.dzangolab.com"
+  },
   "main": "./dist/accordion.common.js",
-  "repository": "github:dzangolab/vue-accordion"
+  "files": [
+    "*.js",
+    "*.json",
+    "*.ts",
+    "dist/*",
+    "lib/*",
+    "public/*"
+  ],
+  "scripts": {
+    "build": "vue-cli-service build --target lib --name accordion ./lib/index.js",
+    "commitlint": "commitlint --edit $1",
+    "lint": "vue-cli-service lint",
+    "prepare": "husky install",
+    "prepublishOnly": "npx sort-package-json && yarn run lint && yarn run test && yarn run build",
+    "prerelease": "yarn run test",
+    "release": "shipjs prepare",
+    "serve": "vue-cli-service serve",
+    "sort-package": "npx sort-package-json",
+    "test": "yarn run test:unit",
+    "test:unit": "vue-cli-service test:unit"
+  },
+  "devDependencies": {
+    "@commitlint/cli": "16.0.1",
+    "@commitlint/config-conventional": "16.0.0",
+    "@fortawesome/fontawesome-svg-core": "1.2.36",
+    "@fortawesome/free-solid-svg-icons": "5.15.4",
+    "@fortawesome/vue-fontawesome": "3.0.0-5",
+    "@vue/cli-plugin-babel": "4.5.15",
+    "@vue/cli-plugin-eslint": "4.5.15",
+    "@vue/cli-plugin-unit-jest": "4.5.15",
+    "@vue/cli-service": "4.5.15",
+    "@vue/compiler-sfc": "3.2.26",
+    "@vue/test-utils": "2.0.0-rc.18",
+    "babel-eslint": "10.1.0",
+    "babel-jest": "26.6.3",
+    "bootstrap": "5.1.3",
+    "core-js": "3.20.2",
+    "eslint": "^7.32, < 8.0",
+    "eslint-plugin-import": "2.25.4",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "6.0.0",
+    "eslint-plugin-vue": "8.2.0",
+    "husky": "7.0.4",
+    "sass": "^1.47.0",
+    "sass-loader": "10.2.0",
+    "shipjs": "0.24.1",
+    "vue": "3.2.26",
+    "vue-jest": "5.0.0-alpha.10"
+  },
+  "peerDependencies": {
+    "vue": ">= 3.0"
+  }
 }
