@@ -1,13 +1,9 @@
-import Vue from 'vue'
+import {createApp} from 'vue'
 import {FontAwesomeIcon} from '@fortawesome/vue-fontawesome'
 import App from './App.vue'
 
 import '@/assets/scss/app.scss'
 
-Vue.component('FontAwesomeIcon', FontAwesomeIcon)
-
-Vue.config.productionTip = false
-
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+createApp(App)
+  .component('font-awesome-icon', FontAwesomeIcon)
+  .mount('#app')
