@@ -97,7 +97,8 @@
 <script>
 import {library} from '@fortawesome/fontawesome-svg-core'
 import {faChevronDown, faChevronUp} from '@fortawesome/free-solid-svg-icons'
-import Accordion from './components/Accordion/Index'
+import Accordion from '../lib/'
+import '../lib/assets/accordion.scss'
 
 library.add(faChevronDown, faChevronUp)
 
