@@ -1,46 +1,44 @@
-import {shallowMount} from '@vue/test-utils'
-import AccordionItem from '@/components/Accordion/Item'
+import {mount} from '@vue/test-utils'
+import AccordionItem from '@/components/Item'
 
 describe('AccordionItem', () => {
   const title = 'title'
   const content = 'content'
   const index = 0
 
-  const wrapperFactory = ({propsData, scopedSlots} = {}) => {
-    return shallowMount(AccordionItem, {
-      propsData: {
+  const wrapperFactory = ({props, slots} = {}) => {
+    return mount(AccordionItem, {
+      props: {
         content,
         index,
         title,
-        ...propsData
+        ...props
       },
-      scopedSlots: {
-        ...scopedSlots
-      }
+      slots: slots
     })
   }
 
   it('renders title and content', async () => {
     const wrapper = wrapperFactory()
 
-    expect(await wrapper.find('[data-el="title"]').text()).toEqual(title)
-    expect(await wrapper.find('[data-el="content"]').text()).toEqual(content)
+    expect(wrapper.find('[data-el="title"]').text()).toEqual(title)
+    expect(wrapper.find('[data-el="content"]').text()).toEqual(content)
 
-    wrapper.destroy()
+    wrapper.unmount()
   })
 
   it('does not render toggle button', async () => {
     const wrapper = wrapperFactory({
-      propsData: {
+      props: {
         showToggle: false
       }
     })
 
-    const toggle = await wrapper.find('[data-el="toggle"]')
+    const toggle = wrapper.find('[data-el="toggle"]')
 
     expect(toggle.exists()).toBe(false)
 
-    wrapper.destroy()
+    wrapper.unmount()
   })
 
   it('emits accordion:select event on click accordion item', async () => {
@@ -54,34 +52,34 @@ describe('AccordionItem', () => {
 
     expect(event[0]).toEqual([index])
 
-    wrapper.destroy()
+    wrapper.unmount()
   })
 
-  it('renders passed value in title after slot passed', async () => {
+  it('renders title slot', async () => {
     const wrapper = wrapperFactory({
-      scopedSlots: {
-        title: `<template><h3>${title}</h3></template>`
+      slots: {
+        title: `<template v-slot:title><h3>${title}</h3></template>`
       }
     })
 
-    const itemTitle = await wrapper.find('[data-el="title"]')
+    const itemTitle = wrapper.find('[data-el="title"]')
 
     expect(itemTitle.html()).toContain(`<h3>${title}</h3>`)
 
-    wrapper.destroy()
+    wrapper.unmount()
   })
 
-  it('renders passed value in content after slot passed', async () => {
+  it('renders content slot', async () => {
     const wrapper = wrapperFactory({
-      scopedSlots: {
-        content: `<template><pre>${content}</pre></template>`
+      slots: {
+        content: `<template v-slot:content><pre>${content}</pre></template>`
       }
     })
 
-    const itemContent = await wrapper.find('[data-el="content"]')
+    const itemContent = wrapper.find('[data-el="content"]')
 
     expect(itemContent.html()).toContain(`<pre>${content}</pre>`)
 
-    wrapper.destroy()
+    wrapper.unmount()
   })
 })
